chore: init FSD
