+++ apps/memory/app/layout.tsx
@@ -1,18 +1,2 @@
-import './global.css';
-
-export const metadata = {
-  title: 'Welcome to memory',
-  description: 'Generated by create-nx-workspace',
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en">
-      <body>{children}</body>
-    </html>
-  );
-}
+import MainLayout from '~/memory/app/layout/main-layout';
+export default MainLayout;
