--- apps/memory/app/page.module.css
@@ -1,2 +1,0 @@
-.page {
-}
