+++ apps/memory/src/1-app/layout/main-layout.tsx
@@ -1,0 +1,16 @@
+export const metadata = {
+  title: 'Welcome to memory',
+  description: 'Generated by create-nx-workspace',
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <body>{children}</body>
+    </html>
+  );
+}
